requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.3
django-multiselectfield==0.1.12
idna==2.10
joblib==0.17.0
nltk==3.5
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.54.1
urllib3==1.26.2
